Refactor folder structure and build
