+++ .gitignore
@@ -1,3 +1,3 @@
 azbrowse.db
 /bin
-azbrowse
+/azbrowse
