+++ cmd/azbrowse/demomode.go
