+++ internal/pkg/handlers/subscription.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/lawrencegripper/azbrowse/armclient"
+	"github.com/lawrencegripper/azbrowse/pkg/armclient"
 )
 
 // SubscriptionExpander expands RGs under a subscription
