+++ pkg/armclient/authutils.go
