+++ pkg/armclient/uuid.go
