+++ pkg/endpoints/endpointinfo_test.go
