--- scripts/ci.sh
@@ -1,4 +1,0 @@
-#/bin/bash
-set -e
-
-docker run -it -e BUILD_NUMBER=$TRAVIS_BUILD_NUMBER -v $PWD:/go/src/github.com/lawrencegripper/azbrowse golang:1.10 bash -f /go/src/github.com/lawrencegripper/azbrowse/scripts/build.sh
